pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mirage"
version = "0.1.0"
description = "A stateful API mock server for integration testing"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",      # for TestClient in tests
]

[project.scripts]
mirage = "mirage.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["mirage"]

[tool.pytest.ini_options]
asyncio_mode = "strict"

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
